Move PartialJsonEqual to testutils.
The function PartialJsonEqual from gamestatejson_tests can be used in
other places as well, e.g. for verifying the expected JSON state of
the upcoming pending-move tracker.  Thus we can move it (together with
its own tests) to the testutils.
