+++ src/gamestatejson_tests.cpp
@@ -20,6 +20,7 @@
 
 #include "prospecting.hpp"
 #include "protoutils.hpp"
+#include "testutils.hpp"
 
 #include "database/account.hpp"
 #include "database/character.hpp"
@@ -40,205 +41,6 @@ namespace pxd
 {
 namespace
 {
-
-/* ************************************************************************** */
-
-/**
- * Checks for "partial equality" of the given JSON values.  This means that
- * keys not present in the expected value (if it is an object) are not checked
- * in the actual value at all.  If keys have a value of null in expected,
- * then they must not be there in actual at all.
- */
-bool
-PartialJsonEqual (const Json::Value& actual, const Json::Value& expected)
-{
-  if (!expected.isObject () && !expected.isArray ())
-    {
-      /* Special case:  If both values are integers, then we compare them
-         explicitly here.  This allows values of type "unsigned int" to be
-         equal to values of type "int" (from golden data).  */
-      if (actual.isInt64 () && expected.isInt64 ()
-            && actual.asInt64 () == expected.asInt64 ())
-        return true;
-
-      if (actual == expected)
-        return true;
-
-      LOG (ERROR)
-          << "Actual value:\n" << actual
-          << "\nis not equal to expected:\n" << expected;
-      return false;
-    }
-
-  if (expected.isArray ())
-    {
-      if (!actual.isArray ())
-        {
-          LOG (ERROR) << "Expected value is array, actual not: " << actual;
-          return false;
-        }
-
-      if (actual.size () != expected.size ())
-        {
-          LOG (ERROR)
-              << "Array sizes do not match: got " << actual.size ()
-              << ", want " << expected.size ();
-          return false;
-        }
-
-      for (unsigned i = 0; i < expected.size (); ++i)
-        if (!PartialJsonEqual (actual[i], expected[i]))
-          return false;
-
-      return true;
-    }
-
-  if (!actual.isObject ())
-    {
-      LOG (ERROR) << "Expected value is object, actual not: " << actual;
-      return false;
-    }
-
-  for (const auto& expectedKey : expected.getMemberNames ())
-    {
-      const auto& expectedVal = expected[expectedKey];
-      if (expectedVal.isNull ())
-        {
-          if (actual.isMember (expectedKey))
-            {
-              LOG (ERROR)
-                  << "Actual has member expected to be not there: "
-                  << expectedKey;
-              return false;
-            }
-          continue;
-        }
-
-      if (!actual.isMember (expectedKey))
-        {
-          LOG (ERROR)
-              << "Actual does not have expected member: " << expectedKey;
-          return false;
-        }
-
-      if (!PartialJsonEqual (actual[expectedKey], expected[expectedKey]))
-        return false;
-    }
-
-  return true;
-}
-
-class PartialJsonEqualTests : public testing::Test
-{
-
-protected:
-
-  bool
-  PartialStrEqual (const std::string& actualStr, const std::string& expectedStr)
-  {
-    Json::Value actual;
-    std::istringstream in1(actualStr);
-    in1 >> actual;
-
-    Json::Value expected;
-    std::istringstream in2(expectedStr);
-    in2 >> expected;
-
-    return PartialJsonEqual (actual, expected);
-  }
-
-};
-
-TEST_F (PartialJsonEqualTests, BasicValues)
-{
-  EXPECT_TRUE (PartialStrEqual ("42", "42"));
-  EXPECT_TRUE (PartialStrEqual ("true", "true"));
-  EXPECT_TRUE (PartialStrEqual ("-5.5", "-5.5"));
-  EXPECT_TRUE (PartialStrEqual ("\"foo\"", " \"foo\""));
-
-  EXPECT_FALSE (PartialStrEqual ("42", "0"));
-  EXPECT_FALSE (PartialStrEqual ("1", "1.1"));
-  EXPECT_FALSE (PartialStrEqual ("\"a\"", "\"b\""));
-  EXPECT_FALSE (PartialStrEqual ("true", "false"));
-}
-
-TEST_F (PartialJsonEqualTests, Objects)
-{
-  EXPECT_FALSE (PartialStrEqual ("{}", "5"));
-  EXPECT_FALSE (PartialStrEqual ("5", "{}"));
-
-  EXPECT_FALSE (PartialStrEqual ("{}", R"({"foo": 42})"));
-  EXPECT_TRUE (PartialStrEqual (R"({"foo": 42}")", "{}"));
-
-  EXPECT_TRUE (PartialStrEqual (R"(
-    {"foo": 5, "bar": 42, "baz": "abc"}
-  )", R"(
-    {"bar": 42, "baz": "abc", "test": null}
-  )"));
-
-  EXPECT_FALSE (PartialStrEqual (R"(
-    {"foo": 5}
-  )", R"(
-    {"foo": null}
-  )"));
-  EXPECT_FALSE (PartialStrEqual (R"(
-    {"foo": 5}
-  )", R"(
-    {"foo": 42}
-  )"));
-}
-
-TEST_F (PartialJsonEqualTests, Arrays)
-{
-  EXPECT_FALSE (PartialStrEqual ("[]", "5"));
-  EXPECT_FALSE (PartialStrEqual ("5", "[]"));
-
-  EXPECT_FALSE (PartialStrEqual ("[]", "[5]"));
-  EXPECT_FALSE (PartialStrEqual ("[5]", "[]"));
-  EXPECT_FALSE (PartialStrEqual ("[5]", "[true]"));
-
-  EXPECT_TRUE (PartialStrEqual ("[]", "[]"));
-  EXPECT_TRUE (PartialStrEqual ("[5, -2.5, false]", "[5, -2.5, false]"));
-}
-
-TEST_F (PartialJsonEqualTests, Nested)
-{
-  EXPECT_TRUE (PartialStrEqual (R"(
-    {
-      "foo": [
-        {"abc": 5, "def": 3},
-        {}
-      ],
-      "bar": {
-        "test": [42]
-      }
-    }
-  )", R"(
-    {
-      "foo": [
-        {"abc": 5},
-        {}
-      ],
-      "bar": {
-        "test": [42]
-      }
-    }
-  )"));
-
-  EXPECT_FALSE (PartialStrEqual (R"(
-    {
-      "foo": [
-        {"abc": 5}
-      ]
-    }
-  )", R"(
-    {
-      "foo": [
-        {"abc": null}
-      ]
-    }
-  )"));
-}
 
 /* ************************************************************************** */
